// src/interfaces/DalleTypes.tsx
// Generated by https://quicktype.io

export interface DalleTypes {
    created:   number;
    data:      Datum[];
    prompt:    string;
    _id:       string;
    creator:   string;
    createdAt: string;
    updatedAt: string;
}

export interface Datum {
    url: string;
}

export interface imagesGenerated {
    url:string;
}